split out autocomplete response and result test cases
